Updating UpdateCallerInformationParameters.EWT type

+++ VirtualQNet/Caller/Parameters/UpdateCallerInformationParameters.cs
@@ -7,7 +7,7 @@ namespace VirtualQNet.Caller
     {
         public IEnumerable<string> Skills { get; set; }
         public ExpandoObject Properties { get; set; }
-        public long? EWT { get; set; }
+        public int? EWT { get; set; }
         public string ServiceState { get; set; }
         public int? WaitTimeWhenUp { get; set; }
         public int? TalkTime { get; set; }
